+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `[[runs]]` field to compositions. See [PR 1516]
 - Add `--run-ids` option during `testground run`. See [PR 1516]
 - Add `--result-file ./output.csv` option during `testground run`. See [PR 1516]
+- Move default `TESTGROUND_HOME` from `~/testgraound` to xdg directory specification. See [PR 1544]
 
 ### Fixed
 - Fix dependencies rewrites in the `exec:go` builder. See [PR 1469]
@@ -24,3 +25,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [PR 1469]: https://github.com/testground/testground/pull/1469
 [PR 1481]: https://github.com/testground/testground/pull/1481
 [PR 1537]: https://github.com/testground/testground/pull/1537
+[PR 1544]: https://github.com/testground/testground/pull/1544